requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alpaca-py==0.43.2
alpaca-trade-api==3.2.0
annotated-types==0.7.0
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
deprecation==2.1.0
frozenlist==1.8.0
idna==3.11
msgpack==1.0.3
multidict==6.7.1
numpy==2.4.2
packaging==26.0
pandas==3.0.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.1
requests==2.32.5
schedule==1.2.2
yfinance>=0.2.40
flask>=3.0.0
six==1.17.0
sseclient-py==1.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==1.26.20
websocket-client==1.9.0
websockets==10.4
yarl==1.22.0
